feat(profile): share one account layout with tab navigation
/profile/keys was a bare page while /profile/settings carried all the account
chrome. The header, avatar column and a daisyUI tabs-box bar now live in a
layout both pages use.

The tabs are navigation, not an ARIA tab pattern: each activates a route
change, there is no tabpanel to point aria-controls at, and both links stay
in the normal tab order. Claiming role="tab" would tell a screen reader that
arrow keys switch panels, which is false. A nav landmark plus aria-current
is the correct pattern, and RouterLink's exact-active handling supplies both
the active class and aria-current, so nothing recomputes what the router
already knows.
